rename network package to net to avoid an issue with the linter
