+++ pkg/collector/corechecks/net/doc.go
@@ -7,4 +7,4 @@
 Package network provides core checks for networking
 
 */
-package network
+package net
