+++ pkg/collector/corechecks/net/network_test.go
@@ -1,4 +1,4 @@
-package network
+package net
 
 import (
 	"testing"
